glibc: Fix BZ#19048 and BZ#19243 [ LIN8-2576 ]
Issue: LIN8-2576

BZ#19048 malloc: arena free list can become cyclic, increasing
contention

When a thread leaves, arena_thread_freeres is called, the malloc
arena associated with the thread is added to the head of free_list,
and free_list set to the arena of the exiting thread.

  A common problem can be described as:
1. thread "t1" uses arena "a"
2. thread "t2" uses arena "a"
3. "t1" exit, making:
        a->next_free = free_list;
        free_list = a;
4. "t2" exits, but since free_list == a, it ends with
   free_list->next_free = free_list;

  When a program has several short lived threads, and most commonly
when there are more threads than arenas, one arena will end up being
used by most threads, causing significant contention.

BZ#19243 malloc: reused_arena can pick an arena on the free list,
leading to an assertion failure and reference count corruption

This was introduced by the fix for bug 19048.

If reused_arena picks an arena on the free list (which means that
arena_get2 did not see it yet, and there was a concurrent thread exit,
updating the free list), it will increase the thread reference count of
the arena.  At this point, this invariant in get_free_list no longer
holds:

821  /* Arenas on the free list are not attached to any thread.  */
822  assert (result->attached_threads == 0);

Therefore, the assert is incorrect.  I am not sure if we should simply
remove it and not change anything, or if we should change reused_arena
to avoid this race condition.

Written-by Florian Weimer.

Patches backported from upstream 2.23.

(LOCAL REV; NOT UPSTREAM) -- sent to oe-core on 2016-01-27.

Signed-off-by: Yuanjie Huang <yuanjie.huang@windriver.com>