STIG: The system default umask in /etc/login.defs must be 077
Issue: LIN8-3167

The umask value influences the permissions assigned to files when they
are created. A misconfigured umask value could result in files with
excessive permissions that can be read and/or written to by unauthorized
users.

Also modify /etc/profile in base-files because the /etc/profile file
overloads the behavior of /etc/login.defs, so if we desire the
functionality provided by setting umask to 077 we should set it in
login.defs to adhere to STIG but also set it in /etc/profile to have the
intended functionality.

(LOCAL REV: NOT UPSTREAM) -- pending

Signed-off-by: Alexandru Moise <alexandru.moise@windriver.com>