oe-core:package_manager extract target arch from TARGET_SYS and add it into package_archs
Issue: LINNT7-183

Since the rpmbuild on the target will use TARGET_SYS as the
'target' to build the source rpm packages, which will extracted
the target arch from TARGET_SYS, such as if TARGET_SYS values
'arm-wrs-linux-gnueabi', the target arch will be 'arm'.

In order to make the rpm pkgs built out on the target compatilbe
with target, extract the arch from TARGET_SYS and add it into
package archs which will be write into the rootfs's /etc/rpm/platform.

Signed-off-by: fli <fupan.li@windriver.com>
Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>