openssl: Security Advisory - openssl - CVE-2016-2106
Issue: LIN8-3423

An integer overflow flaw, leading to a buffer overflow, was
found in the way the EVP_EncryptUpdate() function of OpenSSL
parsed very large amounts of input data. A remote attacker
could use this flaw to crash an application using OpenSSL or,
possibly, execute arbitrary code with the permissions of the
user running that application.

Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>