openssl: Security Advisory - openssl - CVE-2016-2108
Issue: LIN8-3420

A flaw was found in the way OpenSSL encoded certain ASN.1
data structures. An attacker could use this flaw to create a
specially crafted certificate which, when parsed by OpenSSL,
could cause it to crash, or execute arbitrary code using the
permissions of the user running an application compiled
against the OpenSSL library.

Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>