openssl: Security Advisory - openssl - CVE-2016-2179
Issue: LINPUL8-215

It was found that current mechanism of queuing the future messages,
i.e.messages having greater sequence numbers that are to be processed
later, is prone to DoS attack by memory exhaustion, when attacker can
fill up the queue with lots of large messages that are never going to
be used. Only up to 10 messages in the future can be buffered and
queue gets cleared when the connection is closed, thus attacker can
exploit this only with opening many simultaneous connections.

Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>